\name{Ant_data.count}
\alias{Ant_data.count}
\docType{data}
\title{
Ant data with incidence counts
}
\description{
  Species \dQuote{incidence counts} for ant data provided by Longino et al. (2002). The ant data which used malaise traps method is collected in a tropical rain forest of Costa Rica. In the ant data set presented 455 species occurrences are distributed among 62 samples with 103 species observed. 
}
\usage{data(Ant_data.count)}
\format{
  a numerical matrix or a data frame of two columns. The first column is the frequency j=1, 2...; and the second column is incidence counts (Qj). \cr\cr
  The ant data is 
  \tabular{rlllll}{
      \tab [j] \tab [Qj] \cr 
      \tab 1  \tab 39 \cr
      \tab 2  \tab 18 \cr
      \tab 3  \tab 13 \cr
      \tab ...  \tab ... \cr
      \tab 19  \tab 2 \cr
      \tab 20  \tab 1 \cr
      \tab 29  \tab 1 \cr
  }
}
\source{
Longino, J.T., Coddington, J. & Colwell, R.K. (2002) The ant fauna of a tropical rain forest: estimating species richness three different ways. \emph{Ecology}, \bold{83}, 689-702.
}
\examples{
data(Ant_data.count)
}
\keyword{datasets}
