\name{Insects_data}
\alias{Insects_data}
\docType{data}
\title{
Frequency Counts for Insects Data : Day-Time
}
\description{
To illustrate our method, we selected the data sets from Janzen (1973a, b) when he collected tropical foliage insects. The following table gives the frequencies for beetles collected respectively in day-time from the site referred to as "Osa primary-hill, dry season, 1967" in Janzen's paper (1973a).
}
\usage{data(Insects_data)}
\format{
  a numerical matrix or a data frame of two columns.
  \tabular{rlllll}{
      \tab [i]  \tab [fi] \cr
      \tab 1  \tab 59 \cr
      \tab 2  \tab 9 \cr
      \tab 3  \tab 3 \cr
      \tab 4  \tab 2 \cr
      \tab 5  \tab 2 \cr
      \tab 6  \tab 2 \cr
      \tab 11  \tab 1 \cr
    }
}
\examples{
data(Insects_data)
}
\keyword{datasets}
