\name{Seed_data}
\alias{Seed_data}
\docType{data}
\title{
Seed data
}
\description{
The seed-bank data proposed by Colwell and Coddington (1994) which contains 121 standardized soil samples with 34 species observed collected from one-hundred 10 m x 10 m grids in a Costa Rican forest. 
}
\usage{data(Seed_data)
}
\format{
  Only the presence/absence of each species is recorded in the matrix. If the species is presence, it is record 1. On the contrary, it is record 0.
 }
\source{
Colwell, R.K. & Coddington, J. A. (1994) Estimating terrestrial biodiversity through extrapolation. \emph{Philosophical Transactions of the Royal Society of London B - Biological Sciences}, \bold{345}, 101-118.
}
\examples{
data(Seed_data)
}
\keyword{datasets}
