\name{ChaoEntropy-package}
\alias{ChaoEntropy-package}
\docType{package}
\title{
Statistical package "\pkg{ChaoEntropy}"
}
\description{
The purpose of this R package \pkg{ChaoEntropy} is mainly to provide a new Shannon entropy estimator proposed by Chao et al. (2013) for both individual-based (abundance) data and sample-based (incidence) data.
}
\details{
\tabular{ll}{
Package: \tab \pkg{ChaoEntropy}\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-20\cr
License: \tab GPL-3\cr
URL: \tab \url{http://chao.stat.nthu.edu.tw/blog/}
}
functions : \kbd{ChaoEntropy}, \kbd{Count2Abun}, \kbd{Count2Inci}, \kbd{Martix2Inci}
}
\author{
Y.T. Wang ; Y.H. Lee ; K.S. Tseng ; Anne Chao \cr
Maintainer : Anne Chao \email{chao@stat.nthu.edu.tw}
}
\keyword{entropy}
\keyword{Diversity}
